split files inorder to structure new OS color setters
