+++ src/main/kotlin/com/github/nowtilous/projectcolor/TitleBarSetter.kt
@@ -1,0 +1,22 @@
+package com.github.nowtilous.projectcolor
+
+import com.intellij.ide.util.PropertiesComponent
+import com.intellij.openapi.project.Project
+import java.awt.Color
+import java.awt.Component
+
+const val COLOR_SETTING_PATH = "com.github.nowtilous.projectcolor.rgb"
+
+val gColorLockedComponentMap = mutableMapOf<Component, Boolean>()
+val gProjectColorMap = mutableMapOf<Project, Color>()
+
+/**
+ * Set title bar color with given color for given project.
+ */
+fun setTitleBarColor(color: Color, project: Project) {
+
+
+
+    // save color config for persistence
+    PropertiesComponent.getInstance(project).setValue(COLOR_SETTING_PATH, color.rgb, 0)
+}
